Move useful code to a utils module.
