+++ crystal_diffusion/analysis/generator_sample_analysis_utils.py
@@ -1,0 +1,72 @@
+import torch
+from einops import einops
+
+from crystal_diffusion.generators.ode_position_generator import \
+    ExplodingVarianceODEPositionGenerator
+from crystal_diffusion.models.mace_utils import get_adj_matrix
+from crystal_diffusion.models.score_networks.score_network import ScoreNetwork
+from crystal_diffusion.samplers.variance_sampler import NoiseParameters
+
+
+class PartialODEPositionGenerator(ExplodingVarianceODEPositionGenerator):
+    """Partial ODE Position Generator.
+
+    This class is for exploring the nature of the sampling process. It differs from the base class
+    by:
+        1- providing a final time, 'tf', from which diffusion ODE solving starts (tf=1 for the actual problem).
+        2- providing a fixed starting point, initial_relative_coordinates, instead of a random starting point.
+    """
+
+    def __init__(self,
+                 noise_parameters: NoiseParameters,
+                 number_of_atoms: int,
+                 spatial_dimension: int,
+                 sigma_normalized_score_network: ScoreNetwork,
+                 initial_relative_coordinates: torch.Tensor,
+                 record_samples: bool = False,
+                 absolute_solver_tolerance: float = 1.0e-3,
+                 relative_solver_tolerance: float = 1.0e-2,
+                 tf: float = 1.0,
+                 ):
+        """Init method."""
+        super(PartialODEPositionGenerator, self).__init__(noise_parameters,
+                                                          number_of_atoms,
+                                                          spatial_dimension,
+                                                          sigma_normalized_score_network,
+                                                          record_samples,
+                                                          absolute_solver_tolerance,
+                                                          relative_solver_tolerance)
+
+        self.tf = tf
+        assert initial_relative_coordinates.shape[1:] == (number_of_atoms, spatial_dimension), "Inconsistent shape"
+
+        self.initial_relative_coordinates = initial_relative_coordinates
+
+    def initialize(self, number_of_samples: int):
+        """This method must initialize the samples from the fully noised distribution."""
+        assert number_of_samples == self.initial_relative_coordinates.shape[0], "Inconsistent number of samples"
+        return self.initial_relative_coordinates
+
+
+def get_interatomic_distances(cartesian_positions: torch.Tensor,
+                              basis_vectors: torch.Tensor,
+                              radial_cutoff: float = 5.0):
+    """Get Interatomic Distances.
+
+    Args:
+        cartesian_positions :
+        basis_vectors : basis vectors of defining the unit cell.
+        radial_cutoff : neighbors are considered  up to this cutoff.
+
+    Returns:
+        distances : all distances up to cutoff.
+    """
+    shifted_adjacency_matrix, shifts, batch_indices = get_adj_matrix(positions=cartesian_positions,
+                                                                     basis_vectors=basis_vectors,
+                                                                     radial_cutoff=radial_cutoff)
+
+    flat_positions = einops.rearrange(cartesian_positions, "b n d -> (b n) d")
+
+    displacements = flat_positions[shifted_adjacency_matrix[1]] - flat_positions[shifted_adjacency_matrix[0]] + shifts
+    interatomic_distances = torch.linalg.norm(displacements, dim=1)
+    return interatomic_distances
